Upgrade Packages
This commit upgrades the packages to their latest version to fix the critical vulnerabilities. In doing this I've also updated the tests to make sure they run correctly. I've also updated the test that was looking for a link to make sure the start game button component is rendered instead.